qt-x11-opensource-src-4.1.3

+++ tools/designer/src/lib/shared/resourcefile.cpp
@@ -394,7 +394,7 @@ QModelIndex ResourceModel::index(int row, int column,
 {
     QModelIndex result;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid()) {
         if (row < m_resource_file.prefixCount())
@@ -413,7 +413,7 @@ QModelIndex ResourceModel::parent(const QModelIndex &index) const
 {
     QModelIndex result;
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     if (index.isValid() && d != -1)
         result = createIndex(d, 0, -1);
@@ -425,7 +425,7 @@ int ResourceModel::rowCount(const QModelIndex &parent) const
 {
     int result = 0;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid())
         result = m_resource_file.prefixCount();
@@ -444,7 +444,7 @@ bool ResourceModel::hasChildren(const QModelIndex &parent) const
 {
     bool result = false;
 
-    qint64 d = parent.internalId();
+    qint32 d = parent.internalId();
 
     if (!parent.isValid())
         result = m_resource_file.prefixCount() > 0;
@@ -476,7 +476,7 @@ QVariant ResourceModel::data(const QModelIndex &index, int role) const
     if (!index.isValid())
         return QVariant();
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     QVariant result;
 
@@ -519,7 +519,7 @@ void ResourceModel::getItem(const QModelIndex &index, QString &prefix, QString &
     if (!index.isValid())
         return;
 
-    qint64 d = index.internalId();
+    qint32 d = index.internalId();
 
     if (d == -1) {
         prefix = m_resource_file.prefix(index.row());
